requirements.txt
networkx==3.4.2
numpy==2.2.4
pillow==11.1.0
pyglet==2.1.3
PyOpenGL==3.1.9
scipy==1.15.2
trimesh==4.6.5
